requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=9.5.0
numpy>=1.21.0,<2.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
mediapipe==0.10.14
python-dotenv>=1.0.0
flask>=2.3.3
requests>=2.31.0
pyttsx3>=2.90
pymongo>=4.6.0
twilio>=8.0.0 
